ctlseqs: Fix a few typos, and update the XML documentation
I also marked which control sequences *appear* to be supported by the Winodws Terminal; determining which are *actually* supported is a little tricky as supported features may differ depending on what is hosting the terminal widget (Windows Terminal, conhost, vscode, etc), and some control sequences are likely parsed but not implemented (e.g., window resizing). For now this information doesn't appear in the big comparison table as there is already too much in there right now.