Implement xterm alternate screen feature (#231)
This should implement all of xterms alternate screen feature except for DECSET-47 as that conflicts with DECGRCM and I've not decided yet which of the two definitions for this mode I'd like the K95 terminal type to implement.

It has been lightly tested and seems to work OK. Behaviour for all four modes seems to match xterm. Due to the nature of the change, and the significant set of changes to enable paging that this is built on, it could really do with some solid testing.

Closes #231