Fix scrollbar not always being disabled properly
when switching away from page 0. For some reason this didn't seem to happen when switching to the alternate buffer which is odd. Perhaps related to the issue where the horizontal scrollbar would partially come to life for some non-obvious reason. Not sure what is special about page 10 to cause that.