Try to only dirty the vscrn if the cursor is on the visible page
This should speed up changes to background page slightly. I've only bothered making this change to control sequences that are implemented by a VT420 or similar - a few SCO and ANSI control sequences were left alone.

Also fixed the scrollback getting disabled when the cursor was on a different page while PCCM wasoff.