Implement DECDMAC extra semicolon quirk for VT420
VT420 seems to count unexpected semicolons not part of a repetition in the DECDMAC sequence as an error causing it to throw away the entire macro, where as the VT520 just throws away the unexpected semicolons and continues on.

With this change, I think K95s macro implementation should match both the VT420 and VT520. Until a VT420 emulation is added only the VT520 behaviour (extra semicolons ignored, tail calls not on the stack) is exposed however.