Some initial work towards making DECRQCRA less wrong
Right now it should correctly compute the checksum for any region of the display that the host has written to completely. If any cells in the DECRQCRA region are erased, then the checksum will be incorrect as erased cells should not be included.

Additionally, colour information is not currently included in the checksum - that still needs to be done too.

And lastly, the checksum is only valid for the latin1 charset - for other VT420/520 charsets further work will need to be done.