Fix a pair of problems with DECFRA
It wasn't checking that the character specified was within the valid range, and it wasn't translating from the currently selected remote character set to the local one. When in UTF-8, no translation is applied and the entire base multilingual plane is allowed.

Also adjusted DECERA so that it only erases non-erased cells.